+++ include/Events.h
@@ -7,6 +7,7 @@
 
 #pragma once
 
+// AST Utilities includes
 #include "SignalService.h"
 
 namespace astu {
@@ -50,7 +51,7 @@ namespace astu {
      *
      * @ingroup input_group
      */
-    using MouseButtonListener = ISignalListener<MouseButtonEvent>;
+    using IMouseButtonListener = ISignalListener<MouseButtonEvent>;
 
     /**
      * This event represents a keystroke event.
@@ -84,7 +85,39 @@ namespace astu {
      *
      * @ingroup input_group
      */
-    using KeystrokeListener = ISignalListener<KeystrokeEvent>;
+    using IKeystrokeListener = ISignalListener<KeystrokeEvent>;
+
+    // Forward declaration
+    class IService;
+
+
+    /**
+     * Service events describe state changes of services, like being started or stopped.
+     */
+    class ServiceEvent {
+    public:
+        enum Type {Started, Stopped};
+
+        IService & service;
+        Type type;
+
+        ServiceEvent(Type type, IService & service)
+            : service(service), type(type) {}
+    };
+
+    /** 
+     * Type definition for signal services used to transmit keystroke events.
+     *
+     * @ingroup input_group
+     */
+    using ServiceEventService = SignalService<ServiceEvent>;
+
+    /** 
+     * Type definition for signal listeners which receive keystroke events.
+     *
+     * @ingroup input_group
+     */
+    using IServiceListener = ISignalListener<ServiceEvent>;
 
 
 } // end of namespace